fix(merge): merge conflicts

+++ x/poa/keeper/v2nRewardCalculator.go
@@ -48,7 +48,7 @@ func (k msgServer) V2NRewardCalculator(ctx sdk.Context, rewardMultiplier float64
 
 	// Protection against +Inf netEarnings calculation
 	if totalMultipliers > 0 {
-		return (rewardMultiplier / totalMultipliers) * rewardPerBlock, nil
+		return (rewardMultiplier / totalMultipliers) * float64(rewardPerBlock), nil
 	} else {
 		return 0, nil
 	}
